requirements.txt
amqp==2.2.2
astroid==1.5.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
ddt==1.1.1
dj-database-url==0.4.2
Django==1.11.6
django-celery-beat==1.1.0
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
ephem==3.7.6.0
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
isort==4.2.15
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
pew==1.1.0
pika==0.11.0
pipenv==8.3.2
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
pylint==1.7.4
PySocks==1.6.7
pytz==2017.2
requests==2.18.4
six==1.11.0
twilio==6.8.3
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
wrapt==1.10.11
